backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
langchain
langchain-community
langchain-google-vertexai
langchain-text-splitters
google-cloud-aiplatform
faiss-cpu # or faiss-gpu if you have CUDA
# chromadb # Alternative vector store
pandas
xlrd>=2.0.1
openpyxl
python-docx
python-pptx
pypdf # PyPDF2 is an alias, pypdf is the successor
markdown
requests
python-dotenv
# For Tika (optional, more complex setup)
# tika
